Kantare Update (incompatible with kollections from older ver.)
Little big update for Kantare:

Instead of duplicating copies of kantan mapper ad infinitum, there are now 2 mapping decks, and the 'mappers' just store the json data from kantanmapper's export in a text dat and passes this on to the first mapdeck which passes its previous mapping on to the B deck used for fades. This gives a huge performance boost and decrease in savefile size.
The network has been more cleaned up with clear layout and coloring.
Some bugs inherent in Kantan Mapper have been worked around using try: except: clauses in various parts of the original code, so updating to newer versions of td in the future might go slower.